Relocating artifacts
