+++ ofono/pushagent.go
@@ -27,8 +27,8 @@ import (
 	"log"
 	"sync"
 
+	"github.com/ubuntu-phonedations/nuntium/mms"
 	"launchpad.net/go-dbus/v1"
-	"launchpad.net/nuntium/mms"
 )
 
 /*
